changer enreg voiture ancienne méthode

+++ serveur/inventaireVoiture/includes/Voiture.inc.php
@@ -9,7 +9,7 @@ class Voiture {
     private $prix;
     private $quantite;
 
-    function __construct(int $idV, string $nomVoiture, string $description, string $image, float $prix, int $quantite) {
+    function __construct(int $idV, string $nomVoiture, string $description, string $image, int $prix, int $quantite) {
         $this->setIdv($idV);
         $this->setNomVoiture($nomVoiture);
         $this->setDescription($description);
@@ -30,7 +30,7 @@ class Voiture {
     function getImage():string {
         return $this->image;
     }
-    function getPrix():float {
+    function getPrix():int {
         return $this->prix;
     }
     function getQuantite():int {
